Title: Add the Dependency Graph feature
Description: This allows user to check the dependencies and their relationship to a service, with the git details such as description and team owner where available. The `service` feature shows relationship between services while the `system` feature shows the interconnectedness of the entire docker-chs-development services. The `service` feature also enable user to process output through 3 visual method - as a diagram, a tree or a flattree. The `system` feature however process output only as a diagram.