INO-11: Feature -> Add Dependency Graph (#171)
* Title: Add the Dependency Graph feature

Description: This allows user to check the dependencies and their relationship to a service, with the git details such as description and team owner where available. The `service` feature shows relationship between services while the `system` feature shows the interconnectedness of the entire docker-chs-development services. The `service` feature also enable user to process output through 3 visual method - as a diagram, a tree or a flattree. The `system` feature however process output only as a diagram.

* Fix test cases

- make the dependency properties optional in Service definition to avoid multiple test case fails.
- update test cases to match the new error message for missing repository URL.

* revert otel documentation

* Add functional comments, and modify complex logic

* add unit test cases

* fix test case

* add clickable Git repo links to diagram nodes

Each dependency node in the diagram now includes a link to its Git repository,
allowing users to quickly navigate to the corresponding repo for more details.